style: remove style redundencies between pages
